+++ common/src/main/java/smartin/miapi/item/modular/VisualModularItem.java
@@ -10,6 +10,13 @@ public interface VisualModularItem {
         return isVisualModularItem(itemStack, itemStack.getItem());
     }
 
+    static boolean isModularItemNoComponent(ItemStack itemStack) {
+        return isModularItemNoComponent(itemStack.getItem());
+    }
+
+    static boolean isModularItemNoComponent(Item item) {
+        return item instanceof VisualModularItem;
+    }
     static boolean isVisualModularItem(ItemStack itemStack, Item item) {
         return itemStack.has(ModuleInstance.MODULE_INSTANCE_COMPONENT)
                && item instanceof VisualModularItem;